requirements.txt
readyapi
uvicorn[standard]
motor
elasticsearch
python-dotenv
feedparser
requests
beautifulsoup4
slowapi 